+++ frontend/components/auth/SignUpModal/styled.ts
