+++ frontend/components/common/SkeletonBook/styled.ts
