+++ frontend/pages/_app.tsx
@@ -6,7 +6,7 @@ import { ToastContainer } from 'react-toastify';
 import { RecoilRoot } from 'recoil';
 import { ThemeProvider } from 'styled-components';
 
-import CheckSignInByToken from '@components/CheckSignInByToken';
+import CheckSignInStatus from '@components/auth/CheckSignInStatus';
 import GlobalStyle from '@styles/GlobalStyle';
 import responsive from '@styles/responsive';
 
@@ -15,13 +15,13 @@ import '@styles/font.css';
 export default function App({ Component, pageProps }: AppProps) {
   return (
     <RecoilRoot>
-      <CheckSignInByToken>
+      <CheckSignInStatus>
         <GlobalStyle />
         <ThemeProvider theme={responsive}>
           <Component {...pageProps} />
           <ToastContainer limit={3} />
         </ThemeProvider>
-      </CheckSignInByToken>
+      </CheckSignInStatus>
     </RecoilRoot>
   );
 }
